feat: scheduling functionality for lambda
Create a simple scheduling function for Lambda using the Cloudwatch
Event, with each event corresponded to a config. See variables.tf
for an example of config.

These events are disabled by default so that we can debug a bit before
enabling them in AWS Console. This can be removed later.